api/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pandas==2.0.3
python-dotenv==1.0.0
psycopg2-binary==2.9.9
scikit-learn==1.3.0
xgboost==1.7.6